Moved files to utils folder
