+++ backend/src/main/java/io/swagger/utils/RFC3339DateFormat.java
@@ -1,4 +1,4 @@
-package io.swagger;
+package io.swagger.utils;
 
 import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
 import com.fasterxml.jackson.databind.util.ISO8601Utils;
