+++ backend/src/main/resources/application.properties
@@ -1,5 +1,5 @@
 springfox.documentation.swagger.v2.path=/api-docs
 server.contextPath=
 server.port=8081
-spring.jackson.date-format=io.swagger.RFC3339DateFormat
+spring.jackson.date-format=io.swagger.utils.RFC3339DateFormat
 spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
